# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: StayHub V2 API
  version: 1.0.0
  description: Static OpenAPI contract aligned with the currently implemented StayHub V2 controllers.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v2/accommodations:
    post:
      tags:
        - Accommodations
      summary: Create accommodation
      description: |
        Creates a new accommodation listing owned by the authenticated host.

        The backend derives the host from the authenticated JWT user and forces:
        - `available = true`
        - `deleted = false`
      operationId: createAccommodation
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccommodationRequest'
      responses:
        '201':
          description: Accommodation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAccommodationResponse'
        '400':
          description: Validation error or malformed request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Authenticated user is not allowed to create accommodations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - Accommodations
      summary: Search accommodations by city
      description: Returns active and available accommodations matching the requested city.
      operationId: searchAccommodationsByCity
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: city
          required: true
          description: City name used to filter accommodations.
          schema:
            type: string
            minLength: 1
            example: Armenia
        - in: query
          name: page
          required: false
          description: Zero-based page number.
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          required: false
          description: Page size.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Accommodations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccommodationSearchByCityPage'
        '400':
          description: Invalid city, page or size parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v2/accommodations/{id}:
    get:
      tags:
        - Accommodations
      summary: Get accommodation by ID
      description: |
        Returns the accommodation detail for an active, non-deleted accommodation.

        Note:
        The current response model exposes the host through an entity-backed nested object,
        so the documented `host` schema mirrors the actual serialized structure in the current codebase.
      operationId: getAccommodation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Accommodation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Accommodation retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccommodationGetByIdResponse'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Accommodation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Accommodations
      summary: Deactivate accommodation
      description: Soft deletes an accommodation if it has no future active reservations and the authenticated user is its host owner.
      operationId: deactivateAccommodation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          description: Accommodation identifier.
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: Accommodation deactivated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Active future reservations exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized, valid JWT token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Authenticated user is not the owner of the accommodation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Accommodation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      description: Standard error payload returned by the API.
      properties:
        message:
          type: string
          example: Invalid login attempt detected
        code:
          type: integer
          example: 401
      required:
        - message
        - code
    MessageResponse:
      type: object
      description: Generic response with a single message field.
      properties:
        message:
          type: string
          example: Operation completed successfully
      required:
        - message
...
    Role:
      type: string
      enum:
        - GUEST
        - HOST
...
    CreateAccommodationRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
          example: Cabana familiar con vista al valle
        description:
          type: string
          maxLength: 1000
          example: Cabana campestre equipada para familias, con cocina integral, zona BBQ y vista panoramica al valle.
        capacity:
          type: integer
          minimum: 1
          maximum: 50
          example: 6
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: COP
        pricePerNight:
          type: number
          format: double
          minimum: 0.01
          example: 180000.00
        mainImage:
          type: string
          format: uri
          example: https://images.example.com/accommodations/main/cabana-valle.jpg
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -75.6811
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 4.5339
        locationDescription:
          type: string
          maxLength: 100
          example: A 10 minutos del Parque del Cafe, sobre via principal pavimentada.
        city:
          type: string
          maxLength: 100
          example: Armenia
        images:
          type: array
          minItems: 1
          items:
            type: string
            format: uri
            example: https://images.example.com/accommodations/gallery/cabana-valle-sala.jpg
      required:
        - title
        - description
        - capacity
        - currency
        - pricePerNight
        - mainImage
        - longitude
        - latitude
        - locationDescription
        - city
        - images
    CreateAccommodationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 15
        hostId:
          type: integer
          format: int64
          example: 8
        hostEmail:
          type: string
          format: email
          example: host@example.com
        title:
          type: string
          example: Cabana familiar con vista al valle
        description:
          type: string
          example: Cabana campestre equipada para familias, con cocina integral, zona BBQ y vista panoramica al valle.
        capacity:
          type: integer
          example: 6
        currency:
          type: string
          example: COP
        pricePerNight:
          type: number
          format: double
          example: 180000.00
        mainImage:
          type: string
          format: uri
          example: https://images.example.com/accommodations/main/cabana-valle.jpg
        longitude:
          type: number
          format: double
          example: -75.6811
        latitude:
          type: number
          format: double
          example: 4.5339
        locationDescription:
          type: string
          example: A 10 minutos del Parque del Cafe, sobre via principal pavimentada.
        city:
          type: string
          example: Armenia
        images:
          type: array
          items:
            type: string
            format: uri
        available:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: '2026-04-28T20:45:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-28T20:45:00'
      required:
        - id
        - hostId
        - hostEmail
        - title
        - description
        - capacity
        - currency
        - pricePerNight
        - mainImage
        - longitude
        - latitude
        - locationDescription
        - city
        - images
        - available
        - createdAt
        - updatedAt
    User:
      type: object
      description: Current serialized shape of the nested host object returned by the accommodation detail endpoint.
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          format: email
          example: carlos.ramirez@email.com
        password:
          type: string
          example: $2a$10$hashedPasswordValue
        roles:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Role'
        fullName:
          type: string
          example: Carlos Ramirez
        phoneNumber:
          type: string
          example: +573001112233
        birthDate:
          type: string
          format: date
          example: '1988-04-12'
        deleted:
          type: boolean
          example: false
        profilePicture:
          type: string
          format: uri
          nullable: true
          example: https://images.example.com/profiles/carlos.jpg
        passwordRecoveryCode:
          type: string
          nullable: true
          example: null
        passwordRecoveryExpiration:
          type: string
          format: date-time
          nullable: true
          example: null
        createdAt:
          type: string
          format: date-time
          example: '2026-04-01T10:00:00'
        updatedAt:
          type: string
          format: date-time
          example: '2026-04-10T08:30:00'
      required:
        - id
        - email
        - password
        - roles
        - fullName
        - phoneNumber
        - birthDate
        - deleted
        - createdAt
        - updatedAt
    AccommodationGetByIdResponse:
      type: object
      properties:
        host:
          $ref: '#/components/schemas/User'
        title:
          type: string
          maxLength: 100
          example: Beachfront Villa with Pool
        description:
          type: string
          maxLength: 1000
          example: Spacious villa with private pool and direct beach access.
        capacity:
          type: integer
          minimum: 1
          example: 6
        pricePerNight:
          type: number
          format: double
          example: 250000.00
        mainImage:
          type: string
          format: uri
          example: https://images.example.com/accommodations/main/beachfront-villa.jpg
        locationDescription:
          type: string
          maxLength: 100
          example: Near the main beach and local restaurants
        city:
          type: string
          maxLength: 100
          example: Cartagena
        images:
          type: array
          items:
            type: string
            format: uri
        available:
          type: boolean
          example: true
      required:
        - host
        - title
        - description
        - capacity
        - pricePerNight
        - locationDescription
        - city
        - images
        - available
    AccommodationSearchByCityResponse:
      type: object
      description: Compact accommodation data returned by city search.
      properties:
        accommodationCode:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Cabana familiar en Armenia
        city:
          type: string
          example: Armenia
        capacity:
          type: integer
          example: 5
        pricePerNight:
          type: number
          format: double
          example: 180000.00
        currency:
          type: string
          example: COP
        mainImage:
          type: string
          format: uri
          example: https://images.example.com/accommodations/main/cabana-armenia.jpg
      required:
        - accommodationCode
        - title
        - city
        - capacity
        - pricePerNight
        - currency
        - mainImage
    AccommodationSearchByCityPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AccommodationSearchByCityResponse'
        pageable:
          type: object
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        numberOfElements:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
      required:
        - content
...